# [FIX] point_of_sale: speed up reading computed session fields