@Yenthe666 2021-07-09
God it is nice to have you at Odoo @yelizariev ❤️